test(044): split tx-diff spec
Keep Cardano.Node.Client.TxDiffSpec as the unit-test aggregator, and move generic open-value coverage and Conway fixture coverage into separate TxDiff.CoreSpec and TxDiff.ConwaySpec modules.

Verification:
- nix develop --quiet -c cabal test cardano-node-clients:unit-tests -O0 --test-show-details=direct --test-options '--match TxDiff'
- nix develop --quiet -c just ci
