+++ test/Cardano/Node/Client/TxDiff/ConwaySpec.hs
@@ -1,0 +1,100 @@
+{-# LANGUAGE OverloadedLists #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE TypeApplications #-}
+
+module Cardano.Node.Client.TxDiff.ConwaySpec (spec) where
+
+import Data.Aeson ((.=))
+import Data.Aeson qualified as Aeson
+import Data.Map.Strict qualified as Map
+import Data.Text qualified as Text
+import Lens.Micro ((&), (.~), (^.))
+import Test.Hspec
+
+import Cardano.Ledger.Api.Tx (bodyTxL)
+import Cardano.Ledger.Api.Tx.Body (feeTxBodyL)
+import Cardano.Ledger.Binary (
+    Annotator,
+    Decoder,
+    decCBOR,
+    decodeFullAnnotatorFromHexText,
+    natVersion,
+ )
+import Cardano.Ledger.Coin (Coin (..))
+import Cardano.Node.Client.Ledger (ConwayTx)
+import Cardano.Node.Client.TxDiff (
+    DiffChange (..),
+    DiffNode (..),
+    DiffPath (..),
+    diffConwayTx,
+ )
+
+spec :: Spec
+spec =
+    describe "Conway transactions" $
+        it "reports a Conway fee change at body.fee" $ do
+            tx <- loadFixture sampleHash
+            let tx' = tx & bodyTxL . feeTxBodyL .~ Coin 42
+            diffConwayTx tx tx'
+                `shouldBe` DiffNode
+                    rootPath
+                    ( DiffObject
+                        Map.empty
+                        ( Map.fromList
+                            [
+                                ( "body"
+                                , DiffNode
+                                    (DiffPath ["body"])
+                                    ( DiffObject
+                                        Map.empty
+                                        ( Map.fromList
+                                            [
+                                                ( "fee"
+                                                , DiffNode
+                                                    (DiffPath ["body", "fee"])
+                                                    ( DiffChanged
+                                                        (coinJson (tx ^. bodyTxL . feeTxBodyL))
+                                                        (coinJson (Coin 42))
+                                                    )
+                                                )
+                                            ]
+                                        )
+                                        Map.empty
+                                        Map.empty
+                                    )
+                                )
+                            ]
+                        )
+                        Map.empty
+                        Map.empty
+                    )
+
+rootPath :: DiffPath
+rootPath =
+    DiffPath []
+
+loadFixture :: String -> IO ConwayTx
+loadFixture hash = do
+    hex <- Text.strip . Text.pack <$> readFile (fixturePath hash)
+    case decodeFullAnnotatorFromHexText
+        (natVersion @11)
+        "tx-diff fixture"
+        (decCBOR :: forall s. Decoder s (Annotator ConwayTx))
+        hex of
+        Right tx ->
+            pure tx
+        Left err ->
+            expectationFailure ("failed to decode fixture: " <> show err)
+                >> fail "fixture decode failed"
+
+fixturePath :: String -> FilePath
+fixturePath hash =
+    "test/fixtures/mainnet-txbuild/" <> hash <> ".cbor.hex"
+
+sampleHash :: String
+sampleHash =
+    "789f9a1393e3c9eacd19582ebb1b02b777696c8ddcedda2d8752cb5723c42ef6"
+
+coinJson :: Coin -> Aeson.Value
+coinJson (Coin lovelace) =
+    Aeson.object ["lovelace" .= lovelace]
